+++ frontend/src/hooks/index.ts
@@ -1,3 +1,6 @@
 export { useFormatTableData } from "./useFormatTableData";
 export { useDrawer } from "./useDrawer";
 export { usePickerValues } from "./usePickerValues";
+export { useFormLock } from "./useFormLock";
+export { useFormSubmit } from "./useFormSubmit";
+export { useFormControls } from "./useFormControls";